// src/app/cadastro-base/atributo/models/atributo.model.ts
export class Atributo {
  idAtributo: number;
  id_atributo: number;
  letra: string;
  descricao: string;
  TB_ABRANGENCIA_id_abrangencia: string;
  TB_COMPLEXIDADE_id_complexidade: string;
  TB_IMPACTO_id_impacto: string;
  abrangenciaNome: string;
  complexidadeNome: string;
  impactoNome: string;
}